Move JL parity tests from integration tests to unit tests
All 40 Julia ProblemReductions.jl parity tests (model evaluations and
reduction round-trips) are now in src/unit_tests/jl_parity.rs, removing
duplication with the existing unit test suite. The integration test file
tests/suites/jl_parity.rs is reduced to a stub comment.
